backend/requirements.txt
Flask==2.3.3
Flask_SQLAlchemy==3.0.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
gunicorn==20.1.0
flask-migrate==4.0.4
